pyproject.toml
[project]
name = "css-dev-manager"
version = "0.1.0"
description = "DevManager - Sophisticated installer and auto-updater for DevAutomator Python application"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "pygithub>=1.59.0",
    "pyinstaller>=6.0.0",
    "packaging>=23.0",
    "cryptography>=41.0.0",
    "psutil>=5.9.0",
    "pyjwt>=2.8.0",
    "httpx>=0.28.1",
    "winshell>=0.6; sys_platform == 'win32'",
    "pywin32>=310; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.11.13",
]
windows = [
    "pywin32>=306; sys_platform == 'win32'",
    "winshell>=0.6; sys_platform == 'win32'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
